Allow singleton use of Delux
This simplifies the API by letting you skip passing a pid or GenServer
name to every API call. The reason is that Delux tends to be a singleton
anyway for most uses. In fact, accidentally creating two Delux
GenServers that control the same LEDs was not getting detected unless
names were used.

This is not a backwards compatible change since the Delux GenServer now
has a name when you don't specify one. However, it probably won't break
any code since having more than one Delux GenServer is a uncommon
configuration.