Manually expand default arguments to `render/3` and `info/2`
Since it's easiest to use the default GenServer, this makes sure that
the functions that take fewer arguments prefer the non-GenServer ones.
Using the default way that default arguments are expanded, you need to
specify the GenServer or an argument that's definitely not a pid or
GenServer name will be used for the GenServer. This is a little
confusing when you first see it and less convenient.