Move duration calculation to the end
Due to timing limitations when running patterns, not all durations work
and those that due work require different values to be sent to Linux for
accuracy. This is problematic since the calculations and adjustments for
this need to be handled at the last possible moment. Doing them early
makes other duration calculations complicated.

This moves around a lot of code. The end result is that the compile step
for patterns just needs to provide semantics for what to do when the
pattern ends. This is called `:mode`. It can either loop or be a oneshot
pattern now. This highlighted another complexity which is whether to
restart interrupted patterns at where they should be if time passed or
whether to restart them. If you're blinking LEDs with a regular period,
this doesn't matter, but it could matter under other scenarios. The
`:mode` parameter lets us add that feature when needed.

Another update that happened was that indicators that don't change
aren't updated unnecessarily. The unit tests now reflect this. This is
important since if you're running a Morse code or number blink pattern
on one indicator, changing a second indicator would restart the first
indicator's pattern. That was undesirable and was easy to fix with this
change.