Compensate for the Linux HZ configuration
This adds an `:hz` option to the glue initialization to allow the user
to specify the HZ setting for their Linux kernel. The default is 1000 Hz
which is a very minimal adjustment that will always be very slightly
better than no adjustment. The HZ adjustment is especially needed for
100 Hz kernels and to get the overall duration right for oneshot
patterns.

This fixes Morse code timing at least to the limits of Linux. Delux
won't miscalculate the expected duration of a pattern and cut it off
early.